pyproject.toml
[project]
name = "sequence-annotation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.3",
    "numpy>=2.2.6",
    "pytorch-crf>=0.7.2",
    "scikit-learn>=1.7.2",
    "torch>=2.9.0",
]
